add expand cigar; restructure
