+++ src/common.hpp
@@ -1,0 +1,2 @@
+char get_nuc(int x);
+char get_complement_base(char x);
